Update README.md
The game will be text-driven from a GUI that will accept the following commands:
 
1.	Help (displays this information)
2.	Quit (closes the game)
3.	Move (moves the player into the specified room; the room is then described).
4.	Shoot (fires an arrow into the designated room)
5.	Collect (adds the named item of treasure to your inventory)
6.	Drop (removes the named item of treasure from your inventory – there is a fixed number of items that can be carried)
7.	Inventory (lists your current inventory)
8.	Room (describes the current room)
 
The initial user requirements are for a client/server system implemented using the OCSF framework, with each client interacting with a separate game instance. Game instances will share a common layout which will be loaded from a Java DB database.