Create WumpusMainClass.java
In this game, the nodes (rooms) are attractions, which are serviced by 4 tracks, all originating from the park office. The track structure is not modelled explicitly -it is captured in the edge list. It can also be inferred from the node list - the numbers in the attraction name indicate the track on which the attraction is on.
In the node list below, there is NO duplication of attractions - attractions that are at the junction of two or more tracks are included as comment lines. Note also that there is only one track servicing adjacent attractions.